fixing servo motor error
