+++ PEAT/bg_main.py
@@ -9,7 +9,7 @@ from gps3 import gps3
 import RPi.GPIO as GPIO
 
 from utils.communications import *
-from utils.movement import *
+from PEAT.utils.bg_files.movement import *
 from utils.algaecide import *
 from utils.pins import *
